Access data from user's accelerometers: Server function

Use shiny on your mobile phone to gather accelorameter data for whatever your heart could desire. Note that this has been recently hamstrung by browser's attempts to better secure user data. It still works on mobile devices but motion detection needs to be enabled in the devices settings.

shinymovr(input, output, session, movements = c("x", "y", "z", "gamma",
  "beta", "alpha"), orientations = c("alpha", "beta", "gamma"),
  time_limit = -1, recording_message = "Recording Movement...",
  normalized = TRUE)

Arguments

input
you can ignore these as it is taken care of by shiny::callModule()

output
you can ignore these as it is taken care of by shiny::callModule()

session
you can ignore these as it is taken care of by shiny::callModule()

movements
list of desired movement directions from sensor

orientations
list of desired orientation directions from sensor (good when not used on phone)

time_limit
number of seconds for data gathering, defaults to until button pressed again.

recording_message
text for the button when recording is taking place.

normalized
Do you want the data for each output normalized by this samples mean and standard deviation? Defaults to true.

Value

Reactive variable that will update as new data comes in.

See also

Examples

 movrData <- callModule(shinymovr, 'movr_button')

#> Error in callModule(shinymovr, "movr_button"): could not find function "callModule"