Gather recorded data from UI.

Upon completion of line draw, returns a reactive variable that contains a vector of the y coordinates of what the user has drawn. This also includes the start point specified with `draw_start` This is the server component of shinydrawr. You never directly use this function but instead call it through the shiny function `callModule()`. See the example for how to do this.

shinydrawr(input, output, session, data, draw_start, raw_draw = FALSE,
  x_key = "x", y_key = "y", y_min = NA, y_max = NA)

Arguments

input
you can ignore this as it is taken care of by shiny

output
you can ignore this as it is taken care of by shiny

session
you can ignore this as it is taken care of by shiny

data
dataframe containing data you want to plot in two of its columns

draw_start
position on the x-axis the true data is blocked off and the user is to draw from.

raw_draw
set to true if you want to not draw any line, just let the user draw everything. Auto sets draw_start to begining of data.

x_key
name of the x column.

y_key
name of the y column.

y_min
value of the lowest possible value the user is allowed to draw, defaults to lowest seen in data.

y_max
value of the highest possible value the user is allowed to draw, defaults to highest seen in data.

Examples

if (FALSE) {
 drawChart <- shiny::callModule(shinydrawr,
    "outbreak_stats",
    random_data,
    draw_start = 15,
    x_key = "time",
    y_key = "metric")
 }